Added drawing of shapes
Can now select shapes using the shapes menu and draw them. Cannot draw
by starting in bottom right corner until we add the coordinate function.
Probably won't to clean up the switch statement at some point so it's
not so much duplicate code. Bug: When we select a shape it will try and
draw that shape under the menu.